project/requirements.txt
mmcv_full==1.4.0
mmdet==2.14.0
mmdet3d==0.17.1
mmsegmentation==0.14.1
numba==0.48.0
numpy==1.22.0
Pillow==9.3.0
PyYAML==6.0
timm==0.4.12
torch==1.9.0